==> data_CROUS/data_processed/processing_Juin/1-cleaning_Juin.ipynb <==
# Preprocessing sur les données de Juin 

Ce notebook permet d'avoir une première vue sur les données, et de les préprocesser.

import pandas as pd 
import numpy as np

Regardons ce qu'il y a dans les fichiers de type "resultats choix" 


df = pd.read_csv('data_CROUS/data/Resultats_choix_LieuDeVie_juin.csv', sep = ';', index_col=False, encoding='latin8', na_values=['NSPP', 'rien'])
print(df.columns)
print(df.shape)    
print(df.index)

df

Regardons les tables précédantes pour uniformiser au maximum les colonnes.


df_0 = pd.read_csv('data_CROUS/data/Resultats_choix_21_octobre.csv', sep = ';', index_col=False)
print(df_0.columns) #Index(['Horodateur', 'Numéro de formulaire ', 'ENTREE','Prise en double de l'entrée ', 'PLAT ', 'DESSERT','Prise en double du dessert'],dtype='object')
print(df_0.shape)    
print(df_0.index)

## Plats

### Gestion des colonnes annexes

print(df['Hors etudiant'].unique())
print(df['Autre '].unique())
#print(df['Autre Ã©ventuellement'].unique())

df.loc[np.where(df['Hors etudiant'] == 'Pizza')[0],['Plat 2']] = 'Pizza'
print('Pizza OK')

df.loc[np.where(df['Hors etudiant'] == 'Grillades')[0],['Plat 2']] = 'Grillades'
print('Grillades OK')

#for autre in ['Burger', 'Frites salade', 'Burger frites', 'Burger frites ']:
#    df.loc[np.where(df['Autre Ã©ventuellement'] == autre)[0],['Plat 2']] = autre
#    print(autre, ' OK')

for autre in ['En plat : frites ', 'Burger et frites']:
    df.loc[np.where(df['Autre '] == autre)[0],['Plat 2']] = autre
    print(autre, ' OK')
    
for autre in ['Part pizza en entree ', 'Pizza entree']:
    df.loc[np.where(df['Autre '] == autre)[0],['Plat 1']] = autre
    print(autre, ' OK')
    
df.loc[np.where(df['Autre '] == '3 yaourts ')[0],['Plat 3']] = 'yaourt + yaourt + yaourt'
print('3 yaourts  OK')

### Mise en forme des noms des plats

Pour commencer, regardons un peu la distribution des différents plats qui ont été choisis. 

# NB : 
# print(type(df['PLAT '])) # <class 'pandas.core.series.Series'> 
# print(type(df[['PLAT ']])) # <class 'pandas.core.frame.DataFrame'>

print(len(df['Plat 2'].unique())) # Combien de plats possibles ? --> 11 plats différents 
print(df['Plat 2'].value_counts()) # Quel sont-ils ? 
df['Plat 2'].unique()

On observe qu'une mise en forme va être nécéssaire. En effet, dans les fichier d'Octobre le plat est séparé en différents élements (viande, féculents, légumes).

#df['Plat 2'] = df['Plat 2'].replace('Chipo semoule', 'Chipo + semoule', regex = True)
#df['Plat 2'] = df['Plat 2'].replace('Cabillaud semoule', 'Cabillaud + semoule', regex = True)
#df['Plat 2'] = df['Plat 2'].replace('Burger frites', 'Burger + frites', regex = True)
#df['Plat 2'] = df['Plat 2'].replace('Burger frites ', 'Burger + frites', regex = True)
#df['Plat 2'] = df['Plat 2'].replace('Frites salade', 'Frites + salade', regex = True)

df['Plat 2'] = df['Plat 2'].replace('En plat : frites ', 'frites', regex = True)
df['Plat 2'] = df['Plat 2'].replace('Burger et frites', 'Burger + frites', regex = True)
df['Plat 2'] = df['Plat 2'].replace(['\(', '\)'], '', regex = True)

df['Plat 2'].unique()

Pour le **fichier Lieu De Vie** (LDV) on a une colonne supplémentaire `garniture`.

df['garniture'].unique()

df['garniture'] = df['garniture'].replace('Les 2', 'pomme de terre + poelee meridionale', regex = True)

df['Plat 2'][df['garniture'].notna()] = df['Plat 2'][df['garniture'].notna()] + ' + ' + df['garniture'][df['garniture'].notna()]

df['Plat 2'].unique()

Puis pour standardiser les données nous enlevons les espaces et les majuscules.

# Enlevons tous les espaces:
df['Plat 2'] = df['Plat 2'].replace(' ', '', regex = True)
print(df['Plat 2'])

# On obtient alors 
print(len(df['Plat 2'].unique())) # Combien de plats possibles ? --> 13 plats differents 
print(df['Plat 2'].value_counts()) # Quel sont-ils ? 

df['Plat 2'].unique()

De plus, on observe qu'à cause d'un problème de casse, "FeuilletéSaumonoseille+Torsades+Poeléebrocolis " et "FeuilletéSaumonoseille+Torsades+PoeléeBrocolis" sont considérés comme 2 plats differents 

# --> on enlève la casse 
df['Plat 2'] = df['Plat 2'].str.lower()
# On recommence et on obtient alors :
print(len(df['Plat 2'].unique())) 
print(df['Plat 2'].value_counts()) 
df['Plat 2'].unique()

Les lignes où il n'y a pratiquement que des valeurs anquantes sont à supprimer :

df.loc[np.where(df['Plat 2'].isnull())[0],:]

np.where(df['Plat 2'].isnull() * df['Genre'].isnull())[0]

df = df.drop(index = np.where(df['Plat 2'].isnull() * df['Genre'].isnull())[0])

df.reset_index(drop = True, inplace = True)
df.loc[np.where(df['Plat 2'].isnull())[0],:]

## Entrées

Voyons maintenant pour la colonne "ENTREE"

print(len(df['Plat 1'].unique())) 
print(df['Plat 1'].value_counts()) 
df['Plat 1'].unique()

df['Plat 1'] = df['Plat 1'].replace('Part pizza en entree ', 'pizzaentree', regex = True)
df['Plat 1'] = df['Plat 1'].replace('Pizza entree', 'pizzaentree', regex = True)
df['Plat 1'].unique()

Pas d'erreurs en particulier...  
Nous allons tout de même suppprimer les espaces et mettre en lower case pour standaradiser avec les autres fichiers.

# Enlevons tous les espaces et les apostrophes 
df['Plat 1'] = df['Plat 1'].replace(' ', '', regex = True)
df['Plat 1'] = df['Plat 1'].replace("'", '', regex = True)
# --> on enlève la casse 
df['Plat 1'] = df['Plat 1'].str.lower()
print(len(df['Plat 1'].unique())) 
print(df['Plat 1'].value_counts()) 

## Desserts

Voyons maintenant pour la colonne 'DESSERTS'

print(len(df['Plat 3'].unique())) 
print(df['Plat 3'].value_counts())

# Enlevons tous les espaces et les apostrophes 
df['Plat 3'] = df['Plat 3'].replace(' ', '', regex = True)
df['Plat 3'] = df['Plat 3'].replace("'", '', regex = True)

df['Plat 3'] = df['Plat 3'].replace("\/", '', regex = True)
# --> on enlève la casse 
df['Plat 3'] = df['Plat 3'].str.lower()


On y voit plus clair --> S'il y a des erreurs de frappes / d'enregistrements, on sera plus à meme de le voir.  

print(len(df['Plat 3'].unique())) 
print(df['Plat 3'].value_counts())

# Gestion des colonnes

Des horaires sont écrits dans la colonne Autre pour fichier **ENS**.

#HD = df.loc[np.where(df['Autre Ã©ventuellement'].notna())[0],['Heure départ', 'Autre Ã©ventuellement']]
#AE = df.loc[np.where(df['Autre Ã©ventuellement'].notna())[0],['Heure départ']]
#HD
df['Autre '].unique()

df = df.drop('Hors etudiant', axis=1)
#df = df.drop('Autre Ã©ventuellement', axis=1)
df = df.drop('Autre ', axis=1)
df = df.drop('garniture', axis=1)
print(df.columns)

#Index(['Horodateur', 'Numéro de formulaire', 'ENTREE', 'PLAT ', 'DESSERT'], dtype='object')
df.columns = ['Horodateur', 'Numéro de formulaire', 'ENTREE', 'PLAT ', 'DESSERT',
       'Heure départ', 'Genre', 'Regime', 'Fréquentation', 'Faim',
       'Connaissances', 'Influence', 'Influence aspect', 'Quantité',
       'Equilibré', 'Aspect', 'Convive 1', 'Convive 2', 'Convive 3',
       'Convive 4', 'Convive 5', 'Convive 6', 'Convive 7', 'Convive 8',
       'Unnamed: 27']
df

# Sauvegarde

Le csv est préprocéssé, nous pouvons l'enregistrer. 

df.to_csv("Resultats_choix_Juin_LDV_cleaned.csv", sep = ';')
