==> data_CROUS/data_processed/processing_Juin/2-split_columns_Juin.ipynb <==
# Split Columns données Juin (ENS)

## Importation des données

import numpy as np
import pandas as pd

filename = "Resultats_choix_Juin_LDV_cleaned.csv"
df = pd.read_csv(filename, sep=';')
#df = pd.read_csv(f'../../data_clean/{filename}_cleaned.csv', sep=';')
df

no_horo = False
if df.columns[1]=='Horodateur':
    horodateur = df.pop('Horodateur')
    no_horo = True

pd.DataFrame.pop

## Split

### Entrées

try:
    df[['ENTREE_1', 'ENTREE_2']] = df['ENTREE'].str.split('+', 1, expand=True)
    df.insert(2, 'ENTREE_1', df.pop('ENTREE_1'))
    df.insert(3, 'ENTREE_2', df.pop('ENTREE_2'))
    df.pop('ENTREE')
    split_entree = True
except ValueError as e:
    df['ENTREE_2'] = np.nan
    df.insert(2, 'ENTREE_1', df.pop('ENTREE'))
    df.insert(3, 'ENTREE_2', df.pop('ENTREE_2'))
    split_entree = False
df

if split_entree:
    try:
        df[['ENTREE_2', 'ENTREE_3']] = df['ENTREE_2'].str.split('+', 1, expand=True)
        df.insert(3, 'ENTREE_2', df.pop('ENTREE_2'))
        df.insert(4, 'ENTREE_3', df.pop('ENTREE_3'))
    except ValueError as e:
        df['ENTREE_3'] = np.nan
        df.insert(4, 'ENTREE_3', df.pop('ENTREE_3'))
else:
    df['ENTREE_3'] = np.nan
    df.insert(4, 'ENTREE_3', df.pop('ENTREE_3'))
df

### Plats

df.columns

try:
    df[['PLAT_1', 'PLAT_2']] = df['PLAT '].str.split('+', 1, expand=True)
    df.insert(6, 'PLAT_1', df.pop('PLAT_1'))
    df.insert(7, 'PLAT_2', df.pop('PLAT_2'))
    df.pop('PLAT ')
    split_plat = True
except ValueError as e:
    df['PLAT_2'] = np.nan
    df.insert(6, 'PLAT_1', df.pop('PLAT '))
    df.insert(7, 'PLAT_2', df.pop('PLAT_2'))
    split_plat = False
df

if split_plat:
    try:
        df[['PLAT_2', 'PLAT_3']] = df['PLAT_2'].str.split('+', 1, expand=True)
        df.insert(6, 'PLAT_2', df.pop('PLAT_2'))
        df.insert(7, 'PLAT_3', df.pop('PLAT_3'))
    except ValueError as e:
        df['PLAT_3'] = np.nan
        df.insert(7, 'PLAT_3', df.pop('PLAT_3'))
else:
    df['PLAT_3'] = np.nan
    df.insert(7, 'PLAT_3', df.pop('PLAT_3'))
df

### Desserts

try:
    df[['DESSERT_1', 'DESSERT_2',]] = df['DESSERT'].str.split('+', 1, expand=True)
    df.insert(8, 'DESSERT_1', df.pop('DESSERT_1'))
    df.insert(9, 'DESSERT_2', df.pop('DESSERT_2'))
    df.pop('DESSERT')
    split_dessert = True
except ValueError as e:
    df['DESSERT_2'] = np.nan
    df.insert(8, 'DESSERT_1', df.pop('DESSERT'))
    df.insert(9, 'DESSERT_2', df.pop('DESSERT_2'))
    split_dessert = False
df.iloc[:6,1:14]

if split_dessert:
    try:
        df[['DESSERT_2', 'DESSERT_3']] = df['DESSERT_2'].str.split('+', 1, expand=True)
        df.insert(9, 'DESSERT_2', df.pop('DESSERT_2'))
        df.insert(10, 'DESSERT_3', df.pop('DESSERT_3'))
    except ValueError as e:
        df['DESSERT_3'] = np.nan
        df.insert(10, 'DESSERT_3', df.pop('DESSERT_3'))
else:
        df['DESSERT_3'] = np.nan
        df.insert(10, 'DESSERT_3', df.pop('DESSERT_3'))
df.iloc[:6,1:14]

if no_horo :
    df.insert(1, 'Horodateur', horodateur)

## Sauvegarde

df.to_csv('Resultats_choix_Juin_LDV_split.csv', index=False, encoding="cp1252")
# encoding pour le OE majuscule
