==> data_CROUS/data_processed/processing_Juin/3-Formatting_Formulaires_Juin.ipynb <==
# Mise en forme de la partie formulaire des données de Juin

## Chargement package et données

import numpy as np
import pandas as pd
import os
import math
import re 
os.getcwd()

df = pd.read_csv("Resultats_choix_Juin_LDV_split.csv", sep=",", decimal =".", encoding="cp1252")
df

df.columns

Comme on s'intéresse aux variables liées au formulaire on va aussi gérer un Data Frame `dff`contenant uniquement les variables explicatives renseignées dans les formulaires (pas les varaibles liées au choix du plateau)

dff = df[['Heure départ', 'Genre', 'Regime',
       'Fréquentation', 'Faim', 'Connaissances', 'Influence',
       'Influence aspect', 'Quantité', 'Equilibré', 'Aspect', 'Convive 1',
       'Convive 2', 'Convive 3', 'Convive 4', 'Convive 5', 'Convive 6',
       'Convive 7', 'Convive 8', 'Unnamed: 27']]

## Suppression des lignes mal remplies

Certaines lignes de donnée formulaires ne sont pas renseignées (on obtient des lignes entières de `NaN`) : on supprime ces lignes. Dans un premier temps, on repère les lignes contenant que des `Nan` dans la partie formulaire (`dff`).

index_with_nan = dff.index[dff.isnull().all(axis=1)]
index_with_nan

#Suppression dans df
df.drop(index=index_with_nan, axis=0, inplace=True)
df.reset_index(drop = True, inplace = True)

dff.drop(index=index_with_nan, axis=0, inplace=True)
dff.reset_index(drop = True, inplace = True)

df

## Gestion des colonnes

On veut que les DataFrames de Juin aient les mêmes colonnes que les DataFrames d'Octobre. Pour cela il faut que les DataFrames aient le même nombre de colonnes, que ces colonnes soient dans le même ordre et aient le même nom. Pour l'instant, on ne se préoccupe pas des colonnes liées au choix, ni de l'horodateur, et numéro de forulaire.

**Colonnes des `Formulaires d'Octobre`:**
```
['Genre', 'Age', 'Taille', 'Poids', 'IMC', 'Activité physique',
       'Régime particulier', 'Végé', 'Fréquence CROUS', 'Equilibre alim',
       'Seule ou en groupe', 'Nb groupe entrée', 'Proximité', 'Connaissance',
       'Parti groupe', 'Influ sociale', 'Percept senso', 'Nb file', 'Faim',
       'Stress', 'Nb convives', 'Ami 1', 'Ami 2', 'Ami 3', 'Ami 4', 'Ami 5',
       'Ami 6', 'Ami 7', 'Ami 8', 'Ami 9', 'Ami 10', 'Ami 11', 'Ami 12',
       'Ami 13', 'Ami 14', 'Ami 15', 'Ami 16', 'Ami 17']
```

print(dff.columns)
dff.head()

### Variables catégorielles binaires

Elles seront par la suite codées en 0 et 1.

#on regarde quelles sont les variables binaires
print("Genre : ", df["Genre"].unique())        #Genre dans fichiers d'Octobre
print("Regime : ", df["Regime"].unique())        #Régime particulier
print("Connaissances : ", df['Connaissances'].unique())        #Connaissance
print("Influence aspect : ", df['Influence aspect'].unique())
print("Influence : ", df['Influence'].unique())        #Non binaire mas Influence sociale

Pour `Influence aspect` on voit un espace après le 'non' qui est comptabilisé comme une valeur différente. Il faut le modifier. Pour le `Genre` il y a des valeurs 'NB' (Non binaire ?) comment les traiter ? Pour le moment on les laisse en NA.

df["Genre"] = df["Genre"].replace({'Homme':1, 'Femme':0, 'NB':np.nan})
df["Regime"] = df["Regime"].replace({'non':0, 'oui':1})
df['Influence aspect'] = df['Influence aspect'].replace('7', np.nan, regex=True)
df['Influence aspect'] = df['Influence aspect'].replace({'non':0, 'oui':1})
df['Influence'] = df['Influence'].replace({'non':1, 'oui':2}) #our correspondre à la variable échelle d'Octobre
df['Connaissances'] = df['Connaissances'].replace({'non':0, 'oui':1})

print("Genre : ", df["Genre"].unique())        
print("Regime : ", df["Regime"].unique())        
print("Connaissances : ", df['Connaissances'].unique())        
print("Influence aspect : ", df['Influence aspect'].unique())
print("Influence : ", df['Influence'].unique())        

### Variables numériques

Pour ces jeux de données toutes les variables umériques sont issues de variables catégrielles issues d'échelles.

### Variables categorielles en échelle

#on regarde quelles sont les variables concernées
print("Fréquentation : ", df['Fréquentation'].unique())        #Fréquence CROUS
print("Faim : ", df['Faim'].unique())        #Faim
print("Quantité : ", df['Quantité'].unique())        
print("Equilibré : ", df['Equilibré'].unique())        #Equilibre alim
print("Aspect : ", df['Aspect'].unique())        #Perception senso
#print("Attente : ", df['Attente'].unique())        #Nb file

#### Fréquentation

#on remplace les valeurs de la colonne par des valeurs numériques
df["Fréquentation"] = df["Fréquentation"].replace('<1', 0)
df["Fréquentation"] = df["Fréquentation"].replace('1à3', 1)
df["Fréquentation"] = df["Fréquentation"].replace('>1', 1)
df["Fréquentation"] = df["Fréquentation"].replace('>3', 2)

print("Fréquentation : ", df['Fréquentation'].unique())

#### Faim

#Changement d'échelle de 0 à 3
df["Faim"] = df["Faim"].replace([1, 2, 3],0)
df["Faim"] = df["Faim"].replace([4, 5],1)
df["Faim"] = df["Faim"].replace([6, 7],2)
df["Faim"] = df["Faim"].replace([8, 9, 10],3)

print("Faim : ", df['Faim'].value_counts())  

#### Equilibre

Dans un premier temps on règle le problème du 'è'. On suppose qu'il s'agit d'un '7' mal tapé sur un clavier AZERTY. 

#df["Equilibre"] = df["Equilibre"].replace('è','7')

Nous changeons l'échelle pour qu'elle soit la même que celle utilisée dans les formulaires d'Octobre. Pour ce faire nous changeons au préalable le type de données de chaines de caractères à entiers.

#Changement du type
df["Equilibré"] = df["Equilibré"].astype(int)

#Changement d'échelle 0 à 4.
df["Equilibré"] = df["Equilibré"].replace([0, 1, 2],0)
df["Equilibré"] = df["Equilibré"].replace([3, 4],1)
df["Equilibré"] = df["Equilibré"].replace([5, 6],2)
df["Equilibré"] = df["Equilibré"].replace([7, 8],3)
df["Equilibré"] = df["Equilibré"].replace([9, 10],4)

print("Equilibré : ", df['Equilibré'].value_counts()) 

#### Attente

#Changement d'échelle de 0 à 2
"""df["Attente"] = df["Attente"].replace([0, 1, 2, 3],0)
df["Attente"] = df["Attente"].replace([4, 5, 6],1)
df["Attente"] = df["Attente"].replace([7, 8, 9, 10],2)"""

#print("Attente : ", df['Attente'].value_counts())

#### Aspect

#Changement d'échelle de 0 à 3
df["Aspect"] = df["Aspect"].replace([1, 2, 3],0)
df["Aspect"] = df["Aspect"].replace([4, 5],1)
df["Aspect"] = df["Aspect"].replace([6, 7],2)
df["Aspect"] = df["Aspect"].replace([8, 9, 10],3)

print("Aspect : ", df['Aspect'].value_counts())

### Variables "Ami" (chaine de caractères)

Renommons les colonnes `Ami` pour qu'elles correspondent aux colonnes des données d'Octobre (Ami 1 à 17 même si certaines colonnes sont vides).

df.columns

df.columns = ['Unnamed: 0', 'Horodateur', 'Numéro de formulaire', 'ENTREE_1',
       'ENTREE_2', 'ENTREE_3', 'PLAT_1', 'PLAT_2', 'PLAT_3', 'DESSERT_1',
       'DESSERT_2', 'DESSERT_3', 'Heure départ', 'Genre', 'Regime',
       'Fréquentation', 'Faim', 'Connaissances', 'Influence',
       'Influence aspect', 'Quantité', 'Equilibré', 'Aspect', 'Ami 1',
       'Ami 2', 'Ami 3', 'Ami 4', 'Ami 5', 'Ami 6',
       'Ami 7', 'Ami 8', 'Ami 9']

Maintenant ajoutons les colonnes `Ami` manquantes.

for i in range(10,18):
    df[f'Ami {i}'] = np.NaN
df.iloc[:,20:]

df['Ami 9'].unique()

## Que faire des NA relous ?

Pour l'instant on ne s'en occupe pas, dans les autres csv on les a remplacées par la médiane.

df = df.replace(pd.NA, np.NaN, regex=True)
df = df.replace('', np.NaN, regex=True)

#median = csv_tot.median(skipna = True)
#len(csv_tot.median(skipna = True))

## Verification du types de données

#df.dtypes

df = df.astype({'Genre':int, 'Regime':int, 'Fréquentation':int, 'Faim':int,
       'Connaissances':int, 'Influence':int, "Aspect":int,
       'Ami 1':str, 'Ami 2':str, 'Ami 3':str, 'Ami 4':str, 'Ami 5':str,
       'Ami 6':str, 'Ami 7':str, 'Ami 8':str, 'Ami 8':str, 'Ami 10':str, 'Ami 11':str, 'Ami 12':str, 
       'Ami 13':str, 'Ami 14':str, 'Ami 15':str, 'Ami 16':str, 'Ami 17':str}, errors = 'ignore')

#df.dtypes

## Réorganisation du DataFrame

Colonnes d'Octobre :
```
['Genre', 'Age', 'Taille', 'Poids', 'IMC', 'Activité physique', 'Régime particulier', 'Végé', 'Fréquence CROUS', 'Equilibre alim', 'Seule ou en groupe', 'Nb groupe entrée', 'Proximité', 'Connaissance', 'Parti groupe', 'Influ sociale', 'Percept senso', 'Nb file', 'Faim', 'Stress', 'Nb convives', 'Ami 1', 'Ami 2', 'Ami 3', 'Ami 4', 'Ami 5', 'Ami 6', 'Ami 7', 'Ami 8', 'Ami 9', 'Ami 10', 'Ami 11', 'Ami 12', 'Ami 13', 'Ami 14', 'Ami 15', 'Ami 16', 'Ami 17']
```

Colonnes actuelles (à changer) pour Juin **ENS**:
```
['Unnamed: 0', 'Numéro de formulaire', 'ENTREE_1', 'ENTREE_2','ENTREE_3', 'PLAT_1', 'PLAT_2', 'PLAT_3', 'DESSERT_1', 'DESSERT_2','DESSERT_3', 'Horodateur', 'Genre', 'Régime', 'Fréquentation', 'Faim', 'Connaissance entourage', 'Influence entourage', 'Influence aspect','Quantité', 'Equilibre', 'Aspect', 'Attente', 'Ami 1', 'Ami 2','Ami 3', 'Ami 4', 'Ami 5', 'Ami 6','Ami 7', 'Ami 8']
```

Colonnes actuelles (à changer) pour Juin **LDV**:
```
['Unnamed: 0', 'Horodateur', 'Numéro de formulaire', 'ENTREE_1',
       'ENTREE_2', 'ENTREE_3', 'PLAT_1', 'PLAT_2', 'PLAT_3', 'DESSERT_1',
       'DESSERT_2', 'DESSERT_3', 'Heure départ', 'Genre', 'Regime',
       'Fréquentation', 'Faim', 'Connaissances', 'Influence',
       'Influence aspect', 'Quantité', 'Equilibré', 'Aspect', 'Ami 1', 'Ami 2',
       'Ami 3', 'Ami 4', 'Ami 5', 'Ami 6',
       'Ami 7', 'Ami 8', 'Ami 9', 'Ami 10', 'Ami 11', 'Ami 12','Ami 13', 'Ami 14', 'Ami 15', 'Ami 16', 'Ami 17']
```

Mise en place des colonnes formulaire.

final = df[['Genre',  'Regime',  'Fréquentation',  'Equilibré', 'Connaissances', 'Influence', 
            'Aspect', 'Faim', 'Ami 1', 'Ami 2', 'Ami 3', 'Ami 4', 'Ami 5', 'Ami 6', 'Ami 7', 'Ami 8', 
            'Ami 9', 'Ami 10', 'Ami 11', 'Ami 12', 'Ami 13', 'Ami 14', 'Ami 15', 'Ami 16', 'Ami 17']]

Renommage des colonnes

final.columns = ['Genre', 'Régime particulier', 'Fréquence CROUS', 'Equilibre alim', 'Connaissance', 'Influ sociale', 
                 'Percept senso', 'Faim', 'Ami 1', 'Ami 2', 'Ami 3', 'Ami 4', 'Ami 5', 'Ami 6', 'Ami 7', 
                 'Ami 8', 'Ami 9', 'Ami 10', 'Ami 11', 'Ami 12', 'Ami 13', 'Ami 14', 'Ami 15', 'Ami 16', 'Ami 17']

Ajout des colonnes d'Octobre non renseignées remplies de NA.

#help(pd.DataFrame.insert)

col_Octobre = ['Genre', 'Age', 'Taille', 'Poids', 'IMC', 'Activité physique', 'Régime particulier', 'Végé', 
               'Fréquence CROUS', 'Equilibre alim', 'Seule ou en groupe', 'Nb groupe entrée', 'Proximité', 
               'Connaissance', 'Parti groupe', 'Influ sociale', 'Percept senso', 'Nb file', 'Faim', 'Stress', 
               'Nb convives', 'Ami 1', 'Ami 2', 'Ami 3', 'Ami 4', 'Ami 5', 'Ami 6', 'Ami 7', 'Ami 8', 'Ami 9', 
               'Ami 10', 'Ami 11', 'Ami 12', 'Ami 13', 'Ami 14', 'Ami 15', 'Ami 16', 'Ami 17']

for i in range(len(col_Octobre)):
    if col_Octobre[i] != final.columns[i]:
        final.insert(i, col_Octobre[i], np.nan)

final

Ajout des colonnes de choix non finalisées.

final.insert(0, 'Numéro de formulaire', df['Numéro de formulaire'])

final = pd.merge(final,df[['ENTREE_1', 'ENTREE_2',
       'ENTREE_3', 'PLAT_1', 'PLAT_2', 'PLAT_3', 'DESSERT_1', 'DESSERT_2',
       'DESSERT_3', 'Horodateur']], left_index=True, right_index=True)

print(final.columns)
final

## Sauvegarde du DataFrame

final.to_csv("Formatted_Formulaires_LDV_Juin.csv", encoding='latin1')
