==> data_CROUS/data_processed/processing_Juin/4-One_hot_Choix_Juin.ipynb <==
import pandas as pd 
import numpy as np
import os 

os.getcwd()
#file_out = 'choix_21_octobre_clean.csv'
#file_in = 
df = pd.read_csv('Formatted_Formulaires_LDV_Juin.csv', sep = ',', index_col=False, encoding='latin-1')
print(df.columns)
print(df.shape) 
df

# les lignes prennent les valeur des numeros de formulaire au lieu d'avoir des numeros (sorte de clé etrangère)

#df_meal = df.drop('Unnamed: 0', axis = 1)
#df_meal = df_meal.drop('Numéro de formulaire', axis = 1)
#df_meal = df_meal.drop('Horodateur',axis=1)
df_meal = df[['ENTREE_1', 'ENTREE_2', 'ENTREE_3', 'PLAT_1', 'PLAT_2', 'PLAT_3', 'DESSERT_1', 'DESSERT_2', 'DESSERT_3']]

print(df_meal.columns)
# on va crer la liste de tous les aliments possibles qui sera nos en-têtes de colonnes 
aliments = []
for colonne in df_meal.columns:
    for item in df_meal[colonne].unique():
        if item not in aliments and str(item) != 'nan':
            aliments.append(item)
#aliments.remove('pasdentrée')
#aliments.remove('pasdedessert')
#aliments.remove(nan)
print(aliments)

for aliment in aliments:
    #print(np.isnan(aliment))
    if type(aliment) != str:
        #if aliment
        aliments.remove(aliment) # pour enlever les objets de type None 
    if type(aliment) == float:
        print('cc')
        print(aliment)
        aliments.remove(aliment)

print(type(aliments[10]))

# on crée maintenant un nouveau df one hot dont les colonnes sont les aliments 
df_oh = pd.DataFrame(columns=aliments)

for i in range(len(df)):
    items1 = df_meal.iloc[i].to_list()
    items = []
    #print(items1)
    for item in items1:
        if str(item) != 'nan':
            items.append(item)
    try:
        items.remove('pasdentrée')
    except:
        pass
    try : 
        items.remove('pasdedessert')
    except:
        pass
    #print(items)
    #on crée une ligne de 0 pour l'individu i 
    df_oh.loc[i] = 0
    #print(df_oh)
    # on place des 1 dans les colonnes des items choisis
    for item in items:
        #print(item)
        df_oh.iloc[i, df_oh.columns.get_loc(item)] = 1

print(df_oh)

#df_oh.insert(0, 'Numéro de formulaire ', df['Numéro de formulaire'])
#df_oh.insert(0, 'Horodateur', df['Horodateur'])

df.drop(['Unnamed: 0', 'ENTREE_1', 'ENTREE_2', 'ENTREE_3', 'PLAT_1', 'PLAT_2', 'PLAT_3',
         'DESSERT_1', 'DESSERT_2', 'DESSERT_3'], axis=1, inplace=True)
final = pd.merge(df, df_oh, left_index=True, right_index=True)
final

final.to_csv('Forms-formatted_Choix-onehot_Juin_LDV.csv', sep = ';', encoding='latin-1')
