==> data_CROUS/data_processed/processing_Juin/5-Merging_Juin-Octobre.ipynb <==
# Merging Juin (1 LDV, 1 ENS) and Octobre (2LDV et 2ENS regroupés)

import numpy as np
import pandas as pd
import os
import math

## Importation des données

octobre_all = pd.read_csv("merged_Forms_Choix.csv", sep=",", encoding='latin-1', decimal='.')
juin_ens = pd.read_csv("Forms-formatted_Choix-onehot_Juin_ENS.csv", sep=";", encoding='latin-1' )
juin_ldv = pd.read_csv("Forms-formatted_Choix-onehot_Juin_LDV.csv", sep=";", encoding='latin-1')

print(juin_ens.shape)
print(juin_ldv.shape)
print(octobre_all.shape)

## Vérification des colonnes

print(juin_ens.columns[36:])
print(juin_ldv.columns[36:])
print(octobre_all.columns[36:])

juin_ens.columns

juin_ens.columns= ['Unnamed: 0', 'Numéro de formulaire', 'Genre', 'Age', 'Taille', 'Poids',
       'IMC', 'Activité physique', 'Régime particulier', 'Végé',
       'Fréquence CROUS', 'Equilibre alim', 'Seule ou en groupe',
       'Nb groupe entrée', 'Proximité', 'Connaissance', 'Parti groupe',
       'Influ sociale', 'Percept senso', 'Nb file', 'Faim', 'Stress',
       'Nb convives', 'Ami 1', 'Ami 2', 'Ami 3', 'Ami 4', 'Ami 5', 'Ami 6',
       'Ami 7', 'Ami 8', 'Ami 9', 'Ami 10', 'Ami 11', 'Ami 12', 'Ami 13',
       'Ami 14', 'Ami 15', 'Ami 16', 'Ami 17', 'Horodateur', 'saladesoja',
       'carottesrapees', 'saladeriz', 'saucisson', 'endives', 'sardines',
       'pavedecabillaud', 'lasagnesvg', 'chipo', 'pizza', 'burger', 'semoule',
       'frites', 'fruit', 'yaourtnature', 'donut', 'crèmedessert', 'emmental',
       'tartefrboisesananas', 'moussechocolat', 'chaource']

juin_ldv.columns= ['Unnamed: 0', 'Numéro de formulaire', 'Genre', 'Age', 'Taille', 'Poids',
       'IMC', 'Activité physique', 'Régime particulier', 'Végé',
       'Fréquence CROUS', 'Equilibre alim', 'Seule ou en groupe',
       'Nb groupe entrée', 'Proximité', 'Connaissance', 'Parti groupe',
       'Influ sociale', 'Percept senso', 'Nb file', 'Faim', 'Stress',
       'Nb convives', 'Ami 1', 'Ami 2', 'Ami 3', 'Ami 4', 'Ami 5', 'Ami 6',
       'Ami 7', 'Ami 8', 'Ami 9', 'Ami 10', 'Ami 11', 'Ami 12', 'Ami 13',
       'Ami 14', 'Ami 15', 'Ami 16', 'Ami 17', 'Horodateur', 'oeufmimosa',
       'betterave', 'tomates', 'pizzaentree',
       'saladepommesdeterre', 'escalopedeporc', 'grillades', 'tartesaumon',
       'frites', 'pizza', 'gnocchispestovege', 'burger', 'pommesdeterrevapeur',
       'poeleemeridionale', 'yaourtnature', 'tartechocolat', 'fruit', 'madeleine',
       'fromage', 'pecheausirop']

Correction de répétitions restées dans **octobre_all**.

octobre_all['yaourtauxfruits'] = octobre_all['yaourtauxfruits'] + octobre_all['yaourtauxfruitsenmorceaux']
#octobre_all['rotilegumestofu'] = octobre_all['rotilegumestofu'] + octobre_all['rotidelegumestofu']
octobre_all =octobre_all.drop('yaourtauxfruitsenmorceaux', axis=1)
#octobre_all =octobre_all.drop('rotidelegumestofu', axis=1)

octobre_all['yaourtauxfruits'].unique()
#juin_ens['saladeriz'].unique()

octobre_all.insert(0, 'Numéro de formulaire', octobre_all.pop('index'))

## Problème des Numéro de formulaires communs Juin ENS et LDV

print(juin_ens['Numéro de formulaire'].unique())
print(juin_ldv['Numéro de formulaire'].unique())

On remarque qu'ils sont numérotés de la même façon donc que 'B11' par exemple est dans les deux dataframes ce qu'on ne souhaite pas. Il faut donc corriger la colonne `Numéro de formulaire` mais aussi les colonnes `Ami`.

for col_num_form in ['Numéro de formulaire', 'Ami 1', 'Ami 2', 'Ami 3', 'Ami 4', 'Ami 5', 'Ami 6',
       'Ami 7', 'Ami 8', 'Ami 9', 'Ami 10', 'Ami 11', 'Ami 12', 'Ami 13',
       'Ami 14', 'Ami 15', 'Ami 16', 'Ami 17']:
    juin_ldv[col_num_form] = juin_ldv[col_num_form].replace(to_replace=r'^A', value='E', regex=True)
    juin_ldv[col_num_form] = juin_ldv[col_num_form].replace(to_replace=r'^B', value='F', regex=True)


print(juin_ens['Numéro de formulaire'].unique())
print(juin_ldv['Numéro de formulaire'].unique())

## Horodateur

print(juin_ldv['Horodateur'])
juin_ldv['Horodateur'] = juin_ldv['Horodateur'].replace(" PM", '', regex=True)
juin_ldv['Horodateur']

## Preparation du merged

octobre_all

#list(set(juin_ldv.columns + juin_ens.columns))

final = octobre_all.copy()
entrees = ['saladesoja','carottesrapees', 'saladeriz', 'saucisson', 'endives', 'sardines'] + ['oeufmimosa', 
           'betteravespommesechalotes', 'tomates', 'pizzaentree','saladepommesdeterre']
plats = ['pavedecabillaud', 'lasagnesvg', 'chipo', 'pizza', 'burger', 'semoule',
       'frites'] + ['escalopedeporc', 'grillades', 'tartesaumon', 'gnocchispestovege', 'pommesdeterrevapeur',
       'poeleemeridionale']
desserts = ['fruit', 'yaourtnature', 'donut', 'crèmedessert', 'emmental',
       'tartefrboisesananas', 'moussechocolat', 'chaource'] + ['tartechocolat', 'madeleine',
       'fromage', 'pecheausirop']


for plat in entrees :
    if plat not in final.columns:
        final.insert(int(np.where(final.columns=='torsades')[0]), plat, np.nan)
        
for plat in plats :
    if plat not in final.columns:
        final.insert(int(np.where(final.columns=='yaourtauxfruits')[0]), plat, np.nan)
        
for plat in desserts :
    if plat not in final.columns:
        final.insert(len(final.columns), plat, np.nan)

print(len(set(juin_ldv.columns[41:]).intersection(juin_ens.columns[41:])))
print(juin_ens.shape)
print(juin_ldv.shape)
print(octobre_all.shape)
final.shape

csv_tot = pd.concat([final, juin_ens, juin_ldv])
csv_tot = csv_tot.set_index("Numéro de formulaire")
csv_tot = csv_tot.drop('Unnamed: 0', axis=1)
print(csv_tot.shape)
csv_tot.head()

## Binarisation des NA

csv_tot.iloc[:,38:] = csv_tot.iloc[:,38:].fillna(0)

## Sauvegarde

csv_tot.to_csv('Merged_Octobre-Juin.csv', sep = ';', encoding='latin-1')
