==> data_CROUS/data_processed/processing_Juin/6-Assoc_rules_Choix-forms_Juin_regroupement.ipynb <==
import os
import pandas as pd
import numpy as np
import math

os.getcwd()

#df = pd.read_csv("merged_Forms_Choix_regroupement.csv", encoding="cp1252", on_bad_lines="skip", sep=";")
df=pd.read_csv("Merged_Octobre-Juin.csv", encoding="latin1", sep=";")

df

L = list(df.columns)
len(L)

df["count_fruits"] = df["orange"] + df["pomme"] + df["kiwi"] + df["compote"] + df["poire"] + df["pruneauxausirop"] + df["banane"] + df["cocktailfruitsexotiques"] + df["fruit"] + df["pecheausirop"]

sup = ["orange","pomme","kiwi", "compote","poire","pruneauxausirop","banane","cocktailfruitsexotiques","fruit", "pecheausirop"]
for s in sup:
       L.remove(s)
len(L)

df["count_fromage"] = df["emmental"] + df["bleudauvergneaop"] + df["roquefort"] + df["cantalaop"] + df["fromage"] + df["mimolette"] + df["chaource"]

sup = ["emmental","bleudauvergneaop","roquefort", "cantalaop","fromage","mimolette", "chaource"]
for s in sup:
       L.remove(s)
len(L)

df["count_saladefeculents"] = df["saladepates"] + df["saladepommesdeterre"] + df["saladeriz"] 

sup = ["saladepates","saladepommesdeterre","saladeriz"]
for s in sup:
       L.remove(s)
len(L)

df["count_crudites"] = df["chou"] + df["saladelentilles"] + df["champignons"] + df["carottesrapees"] + df["artichauts"] + df["coeursdepalmiers"] + df["tomates"] + df["betterave"] + df["saladesoja"] + df["endives"] + df["betteravespommesechalotes"] + df["saladeverte"]

sup = ["chou","saladelentilles","champignons","carottesrapees","artichauts","coeursdepalmiers","tomates","betterave","saladesoja", "endives", "betteravespommesechalotes","saladeverte"]
for s in sup:
       L.remove(s)
len(L)

df["count_charcuterie"] = df["oeufmimosa"] + df["salami"] + df["surimi"] + df["minisandwich"] + df["pâté"] + df["rillettepoisson"] + df["saucisson"] + df["sardines"] + df["pizzaentree"]

sup = ["oeufmimosa","salami","surimi","minisandwich","pâté","rillettepoisson","saucisson","sardines", "pizzaentree"]
for s in sup:
    L.remove(s)
len(L)

df["count_viande"] = df["boeufbourguignon"] + df["kebab"] + df["faux-filet"] + df["pizzamerguezpoivrons"] + df["hamburger"] + df["escalopeviennoise"] + df["filetdepoulet"] + df["pizza"] + df["cordonbleu"] + df["boulettesviande"] + df["pizzakebab"] + df["entrecote"] + df["chipo"] + df["burger"] + df["escalopedeporc"] + df["grillades"]

sup = ["boeufbourguignon","kebab","faux-filet","pizzamerguezpoivrons","hamburger","escalopeviennoise","filetdepoulet","pizza","cordonbleu","boulettesviande","pizzakebab","entrecote", "chipo", "burger", "escalopedeporc", "grillades"]
for s in sup:
    L.remove(s)
len(L)

df["count_poisson"] = df["feuilletesaumonoseille"] + df["pavedecabillaud"] + df["pavedepoissonbordelaise"] + df["poissonpane"] + df["tartesaumon"]

sup = ["feuilletesaumonoseille","pavedecabillaud","pavedepoissonbordelaise","poissonpane","tartesaumon"]
for s in sup:
    L.remove(s)
len(L)

df["count_vege"] = df["risottocourgettes"] + df["pizzaraclette"] + df["tajinevege"] + df["burgervege"] + df["rotilegumestofu"] + df["lasagnesvg"] + df["gnocchispestovege"]

sup = ["risottocourgettes","pizzaraclette","tajinevege","burgervege", "rotilegumestofu", "lasagnesvg", "gnocchispestovege"]
for s in sup:
    L.remove(s)
len(L)

df["count_yaourt"] = df["yaourtauxfruits"] + df["faissellenature"] + df["yaourtnature"] + df["fromageblancfruitsrouges"] + df["petitsuissenature"] + df["fromageblanc"] 

sup = ["yaourtauxfruits","faissellenature","yaourtnature","fromageblancfruitsrouges", "petitsuissenature", "fromageblanc"]
for s in sup:
    L.remove(s)
len(L)

df["count_dessertgourmand"] = df["oeufalaneige"] + df["madeleine"] + df["tarteauxfruits"] + df["parisbrest"] + df["gateaudesemoule"] + df["donut"] + df["tartefrboisesananas"] + df["tartechocolat"] 

sup = ["oeufalaneige","madeleine","tarteauxfruits","parisbrest", "gateaudesemoule", "donut", "tartefrboisesananas", "tartechocolat"]
for s in sup:
    L.remove(s)
len(L)

df["count_cremedessert"] = df["moussenoixdecoco"] + df["liegeois"] + df["cremepraline"] + df["flancchocolat"] + df["crèmedessert"] + df["moussechocolat"]

sup = ["moussenoixdecoco","liegeois","cremepraline","flancchocolat", "crèmedessert","moussechocolat"]
for s in sup:
    L.remove(s)
len(L)

df["count_accompagnement"] = df["torsades"] + df["brocolis"] + df["frites"] + df["ratatouille"] + df["pommesdeterrevapeur"] + df["juliennedelegumes"] + df["riz"] + df["spaghettis"] + df["mousselinedepotiron"] + df["legumes"] + df["semoule"] + df["poeleemeridionale"]

sup = ["torsades","brocolis","frites","ratatouille","pommesdeterrevapeur","juliennedelegumes","riz","spaghettis", "mousselinedepotiron", "legumes", "semoule", "poeleemeridionale"]
for s in sup:
    L.remove(s)
len(L)

L

#print(df.columns.get_loc("onehot_fruits"))
print(df.columns.get_loc("count_fruits"))
#print(df.columns[118])

model_col=pd.read_csv("u-merged_Forms_Choix_regrmt_Hélène.csv", encoding="latin1", sep=";")

model_col.iloc[0,:]

for col in df.columns[df.columns.get_loc("count_fruits"):]:
    df['onehot_' + col[6:]] = np.where(df[col] > 0,1,0)
df

def round_to_nearest_IMC_float(x) : 
    if math.isnan(x):
        return x
    
    modulo = round(x)%5
    #print(x, modulo)
    if modulo > 2.5 :
        to_return = round(x) + (5-modulo)
        return to_return
    
    elif modulo == 0 : 
        return round(x)
    
    else : 
        to_return = round(x) - modulo
        return to_return
    
df["IMC"] = df["IMC"].map(lambda IMC : round_to_nearest_IMC_float(IMC))
print(df["IMC"])

#Réduction en classes des colonnes avec des valeurs très diverses pour allégre le one-hot encoding
def round_to_nearest_IMC(x) : 
    modulo = int(x.split(",")[0])%5
    print("Data : ", int(x.split(",")[0]), modulo)
    if modulo > 2.5 :
        to_return = int(x.split(",")[0]) + (5-modulo)
        print("Return : ", to_return)
        return to_return
    
    elif modulo == 0 : 
        print("Return : ", int(x.split(",")[0]))
        return int(x.split(",")[0])
    
    else : 
        to_return = int(x.split(",")[0]) - modulo
        print("Return : ", to_return)
        return to_return
    
#df["IMC"] = df["IMC"].map(lambda IMC : round_to_nearest_IMC(IMC))
#print(df["IMC"])

df.to_csv("Octbre_Juin_Choix_Forms_Regroup_allCol.csv", sep=';', encoding="latin1", index=False)

col_onehot = list(range(df.columns.get_loc("Ami 17")+1)) + list(range(df.columns.get_loc("onehot_fruits"),df.shape[1]))
smaller = df.iloc[:,col_onehot]
smaller

smaller.to_csv("Octbre_Juin_Choix_Forms_Regroup_onehot.csv", sep=';', encoding="latin1", index=False)
